cache database

+++ pyetl/outils/pack.py
@@ -101,10 +101,17 @@ def cache(mapper):
 
     fich2 = os.path.join(place2, "cache_readers.csv")
     fich3 = os.path.join(place2, "cache_writers.csv")
+
+    place4 = os.path.join(os.path.dirname(__file__), "..", "formats", "db")
+    fich4 = os.path.join(place4, "cache_databases.csv")
     if os.path.isfile(fich1):
         os.remove(fich1)
-        os.remove(fich2)
-        os.remove(fich3)
+        if os.path.isfile(fich2):
+            os.remove(fich2)
+        if os.path.isfile(fich3):
+            os.remove(fich3)
+        if os.path.isfile(fich4):
+            os.remove(fich4)
     else:
         with open(fich1, "w") as f:
             for nommodule in sorted(mapper.modules):
@@ -112,7 +119,7 @@ def cache(mapper):
                     for nom in sorted(mapper.modules[nommodule].commandes):
                         f.write(nom + ";" + nommodule + "\n")
 
-        from pyetl.formats.generic_io import READERS, WRITERS
+        from pyetl.formats.generic_io import READERS, WRITERS, DATABASES
 
         with open(fich2, "w") as f:
             for nom, desc in sorted(READERS.items()):
@@ -120,6 +127,10 @@ def cache(mapper):
                 f.write(nom + ";" + desc.module + "\n")
         with open(fich3, "w") as f:
             for nom, desc in sorted(WRITERS.items()):
+                # print("readers", nom, desc.module, desc)
+                f.write(nom + ";" + desc.module + "\n")
+        with open(fich4, "w") as f:
+            for nom, desc in sorted(DATABASES.items()):
                 # print("readers", nom, desc.module, desc)
                 f.write(nom + ";" + desc.module + "\n")
